Refactor image extraction
