+++ old/extract_images_googlevision_auto.py
